+++ xlmhg/__init__.py
@@ -7,12 +7,10 @@ import pkg_resources
 
 __version__ = pkg_resources.require('xlmhg')[0].version
 
-#if 'READTHEDOCS' not in os.environ or \
-#        os.environ['READTHEDOCS'] != 'True':
-    # this is certainly weird, but allows us to import the xlmhg version
-    # number from "docs/conf.py" without running into other problems
 from .result import mHGResult
-from .test import xlmhg_test, get_xlmhg_test_result, get_xlmhg_O1_bound
+from .test import xlmhg_test, get_xlmhg_test_result
+from .visualize import get_result_figure
 
-__all__ = ['mHGResult', 'xlmhg_test', 'get_xlmhg_test_result',
-           'get_xlmhg_O1_bound']
+__all__ = ['xlmhg_test',
+           'mHGResult', 'get_xlmhg_test_result',
+           'get_result_figure']