+++ xlmhg/test.py
@@ -14,7 +14,7 @@
 # You should have received a copy of the GNU General Public License
 # along with this program. If not, see <http://www.gnu.org/licenses/>.
 
-"""Python front-end for the XL-mHG test."""
+"""Python API for performing XL-mHG tests."""
 
 from __future__ import (absolute_import, division,
                         print_function, unicode_literals)